handle limitHit; separate out SearchStats{Page,Languages} components
