+++ web/src/enterprise/search/stats/SearchStatsPage.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import * as GQL from '../../../../../shared/src/graphql/schema'
 import renderer, { act } from 'react-test-renderer'
 import * as H from 'history'
-import { SearchStatsPage, summarizeSearchResultsStatsLanguages } from './SearchStatsPage'
+import { SearchStatsPage } from './SearchStatsPage'
 import { of } from 'rxjs'
 import { MemoryRouter } from 'react-router'
 
@@ -14,7 +14,7 @@ describe('SearchStatsPage', () => {
                     location={H.createLocation({ pathname: '/stats', search: 'q=abc' })}
                     history={H.createMemoryHistory()}
                     _querySearchResultsStats={() =>
-                        of<GQL.ISearchResultsStats>({
+                        of<GQL.ISearchResultsStats & { limitHit: boolean }>({
                             __typename: 'SearchResultsStats',
                             approximateResultCount: '123',
                             sparkline: [],
@@ -25,6 +25,7 @@ describe('SearchStatsPage', () => {
                                 { __typename: 'LanguageStatistics', name: 'D', totalBytes: 0, totalLines: 5 },
                                 { __typename: 'LanguageStatistics', name: '', totalBytes: 0, totalLines: 35 },
                             ],
+                            limitHit: false,
                         })
                     }
                 />
@@ -33,28 +34,36 @@ describe('SearchStatsPage', () => {
                 createNodeMock: () => ({ parentElement: document.implementation.createHTMLDocument().body }),
             }
         )
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
         act(() => undefined) // wait for _querySearchResultsStats to emit
         expect(component.toJSON()).toMatchSnapshot()
     })
-})
 
-describe('summarizeSearchResultsStats', () => {
-    test('collapses low-ranking entries to Other', () =>
-        expect(
-            summarizeSearchResultsStatsLanguages(
-                [
-                    { __typename: 'LanguageStatistics', name: 'A', totalBytes: 0, totalLines: 100 },
-                    { __typename: 'LanguageStatistics', name: 'B', totalBytes: 0, totalLines: 50 },
-                    { __typename: 'LanguageStatistics', name: 'C', totalBytes: 0, totalLines: 10 },
-                    { __typename: 'LanguageStatistics', name: 'D', totalBytes: 0, totalLines: 5 },
-                    { __typename: 'LanguageStatistics', name: '', totalBytes: 0, totalLines: 35 },
-                ],
-                0.1
-            )
-        ).toEqual([
-            { __typename: 'LanguageStatistics', name: 'A', totalBytes: 0, totalLines: 100 },
-            { __typename: 'LanguageStatistics', name: 'B', totalBytes: 0, totalLines: 50 },
-            { __typename: 'LanguageStatistics', name: '', totalBytes: 0, totalLines: 35 },
-            { __typename: 'LanguageStatistics', name: 'Other', totalBytes: 0, totalLines: 15 },
-        ]))
+    test('limitHit', () => {
+        const component = renderer.create(
+            <MemoryRouter>
+                <SearchStatsPage
+                    location={H.createLocation({ pathname: '/stats', search: 'q=abc' })}
+                    history={H.createMemoryHistory()}
+                    _querySearchResultsStats={() =>
+                        of<GQL.ISearchResultsStats & { limitHit: boolean }>({
+                            __typename: 'SearchResultsStats',
+                            approximateResultCount: '123',
+                            sparkline: [],
+                            languages: [
+                                { __typename: 'LanguageStatistics', name: 'A', totalBytes: 0, totalLines: 100 },
+                            ],
+                            limitHit: true,
+                        })
+                    }
+                />
+            </MemoryRouter>,
+            {
+                createNodeMock: () => ({ parentElement: document.implementation.createHTMLDocument().body }),
+            }
+        )
+        // eslint-disable-next-line @typescript-eslint/no-floating-promises
+        act(() => undefined) // wait for _querySearchResultsStats to emit
+        expect(component.toJSON()).toMatchSnapshot()
+    })
 })
